Merge branch 'main' into SWE_repartition_fix

+++ assimilation_code/modules/io/state_vector_io_mod.f90
@@ -274,7 +274,7 @@ type(file_info_type), intent(in)    :: file_info
 logical,              intent(in)    :: use_time_from_file
 type(time_type),      intent(inout) :: model_time
 
-integer :: dart_index ! where to start in state_ens_handle%copies
+integer(i8) :: dart_index ! where to start in state_ens_handle%copies
 integer :: domain
 type(stage_metadata_type) :: restart_files
 
@@ -312,7 +312,7 @@ subroutine write_restart_direct(state_ens_handle, file_name_handle)
 type(ensemble_type),    intent(inout) :: state_ens_handle
 type(stage_metadata_type), intent(in) :: file_name_handle
 
-integer :: dart_index !< where to start in state_ens_handle%copies
+integer(i8) :: dart_index !< where to start in state_ens_handle%copies
 integer :: domain !< loop index
 
 if ( .not. module_initialized ) call state_vector_io_init() ! to read the namelist
